+++ custom_components/ember_mug/entity.py
@@ -1,4 +1,5 @@
 """Generic Entity Logic for multiple platforms."""
+
 from __future__ import annotations
 
 import logging