+++ custom_components/ember_mug/number.py
@@ -1,4 +1,5 @@
 """Binary Sensor Entity for Ember Mug."""
+
 from __future__ import annotations
 
 import logging
@@ -13,16 +14,15 @@ from homeassistant.components.number import (
 from homeassistant.const import UnitOfTemperature
 from homeassistant.helpers.entity import EntityCategory
 
-from .const import DOMAIN, MAX_TEMP_CELSIUS, MIN_TEMP_CELSIUS
+from .const import MAX_TEMP_CELSIUS, MIN_TEMP_CELSIUS
 from .entity import BaseMugValueEntity
 
 if TYPE_CHECKING:
-    from homeassistant.config_entries import ConfigEntry
     from homeassistant.core import HomeAssistant
     from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
+    from . import EmberMugConfigEntry
     from .coordinator import MugDataUpdateCoordinator
-    from .models import HassMugData
 
 
 _LOGGER = logging.getLogger(__name__)
@@ -73,15 +73,15 @@ class MugTargetTempNumberEntity(MugNumberEntity):
 
 async def async_setup_entry(
     hass: HomeAssistant,
-    entry: ConfigEntry,
+    entry: EmberMugConfigEntry,
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up Number Entities."""
     if entry.entry_id is None:
         raise ValueError("Missing config entry ID")
-    data: HassMugData = hass.data[DOMAIN][entry.entry_id]
+    coordinator = entry.runtime_data
     async_add_entities(
         [
-            MugTargetTempNumberEntity(data.coordinator, "target_temp"),
+            MugTargetTempNumberEntity(coordinator, "target_temp"),
         ],
     )