+++ tests/test_number.py
@@ -1,4 +1,5 @@
 """Test Text entities."""
+
 from __future__ import annotations
 
 from typing import TYPE_CHECKING