+++ tests/test_switch.py
@@ -1,4 +1,5 @@
 """Test Switch entities."""
+
 from __future__ import annotations
 
 from typing import TYPE_CHECKING